Various work
Make the McuControlInterface a reference so it can easily be fetched.
Start work on Radio activity.

The Camera Activity is crashing.  It believes its getting an immersive
callback when it is not.  Since the callback is postDelayed, its
executing after the view has been destroyed and thus null.  Hopefully
fixed by removing callbacks before finish is called, however I believe I
am going to move the Camera to a companion app.  May do it with the
radio activity as well.